backend/pyproject.toml
[tool.poetry]
name = "gestao-lab"
version = "0.1.0"
description = ""
authors = ["LucasGamaV <lucas.pymatos@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.36"
psycopg = {extras = ["binary"], version = "^3.2.4"}
passlib = "^1.7.4"
requests = "^2.32.3"
pdoc = "^15.0.1"
pyproj = "^3.7.0"
pydantic-settings = "^2.4.0"
numpy = "^2.1.0"
python-multipart = "^0.0.9"
pyjwt = "2.9.0"
google-auth = "^2.34.0"
pandas = "^2.2.3"
boto3 = "^1.35.24"
moto = "^5.0.15"
paramiko = "^3.5.0"
botocore = "^1.35.71"
polyline = "^2.0.2"
email-validator = "^2.2.0"
pydantic = { version = "2.8.2", extras = ["email"] }


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
mypy = "^1.11.1"
ruff = "^0.5.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"